x265: for public API, only internal bitdepth is configurable
And when HIGH_BIT_DEPTH is 0, the value is ignored. This change allows the input
and output bit depth fields to be configurable for !HIGH_BIT_DEPTH builds, but
are only allowed to be 8.

When x265 is used as a library, the input and output bitdepths are completely
up to the user. libx265 does not have to know what they are.