api: rename enableAMPRefine to useRectInter
This configurable was mis-named.  It has nothing to do with AMP, it is disabling
the use of rectangular inter prediction modes (which has the side-effect of also
disabling AMP since AMP is not checked if the rectangular mode of similar shape
was not checked).

This configurable should never have been in TComSlice since the flag is not
signaled in the slice header like AMP activation is.