testbench: only perform 0xCDCDCDCDCDCD memsets for debug, runs much faster
It should only be necessary to do this when you are not sure the C primitive
is correct.  Once the C primitive is correct, it should never be required to
initialize the output buffers in order to find mismatch bugs.

--HG--
extra : amend_source : 3b82365e6b26aed786923a941426ea7c906d825c