pixelharness: revert memcmp counts to full block size
I had changed a number of these to only compare the actual block size processed
by the primitive, but this was wrongly checking the first n x m pixels instead
of the n x m pixels in the upper left corner of the output buffer (the stride
of the buffers was always 64).  So it was comparing some of the wrong pixels,
and not all of the proper ones.  Comparing all 64x64 pixels is harmless, and
actually verifies the primitive isn't writing beyond the correct outputs.

--HG--
extra : amend_source : 369adbb035f46ca7c608cda63919bc6c1960dcb8