slicetype: alloc wpScalingParam instance as a struct member
This is a workaround for VC11.  When x265 was compiled for debug targeting Win32
the stack was being reported as corrupted by weightCostLuma().  No other
compiler or build option would report any problems (not even valgrind).  In the
VisualStudio debugger the stack would be obviously garbaged once the function
was entered.  Moving `w` off of the stack makes the VC11 debugger happy again.

--HG--
branch : stable